Refactored during review comments
