+++ src/main/java/com/zufar/onlinestore/payment/enums/PaymentCurrency.java
@@ -1,0 +1,14 @@
+package com.zufar.onlinestore.payment.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum PaymentCurrency {
+    USD("usd");
+
+    private final String currency;
+
+    public static final String CURRENCY = "usd";
+}
