+++ src/main/java/com/zufar/onlinestore/payment/exception/UnexpectedPaymentStatusException.java
@@ -1,15 +1,15 @@
 package com.zufar.onlinestore.payment.exception;
 
-import com.zufar.onlinestore.payment.enums.PaymentConstants;
+import com.zufar.onlinestore.payment.enums.PaymentStatus;
 import lombok.Getter;
 
 @Getter
 public class UnexpectedPaymentStatusException extends RuntimeException {
 
-    private final PaymentConstants paymentConstants;
+    private final PaymentStatus paymentStatus;
 
-    public UnexpectedPaymentStatusException(final PaymentConstants paymentConstants) {
-        super(String.format("Payment status: %s is unexpected.", paymentConstants));
-        this.paymentConstants = paymentConstants;
+    public UnexpectedPaymentStatusException(final PaymentStatus paymentStatus) {
+        super(String.format("Payment status = %s is unexpected.", paymentStatus));
+        this.paymentStatus = paymentStatus;
     }
 }
